requirements.txt
azure-functions
slackclient
requests_oauthlib
